requirements.txt
PyQt5>=5.15.0
baidu-aip>=4.16.10
pdf2image>=1.16.3
openpyxl>=3.1.2
Pillow>=9.5.0
pyinstaller>=5.13.0
chardet>=5.2.0